feat: 🎸 improve metadata application and startup time
