+++ src/server.ts
@@ -1,10 +1,8 @@
 import { ENV } from '@/utils/env';
 import { logger, LOG_LEVEL } from './logger';
-import { tag } from './timer';
 
 export const start = async () => {
   logger.info(`Log level: ${LOG_LEVEL}`);
-  tag('Start');
 
   /**
    * Async imports allow to dynamically import the required modules,
@@ -17,22 +15,14 @@ export const start = async () => {
   //   logger.info(`Skipping migrations and metadata`);
   // } else {
   const { waitForHasura } = await import('@/utils');
-  tag('Import waitfor module');
   const { applyMigrations } = await import('./migrations');
-  tag('Import migrations module');
-  // const { applyMetadata } = await import('./metadata');
-  const { patchMetadata } = await import('./metadata-patch');
-  tag('Import metadata module');
+  const { applyMetadata } = await import('./metadata');
 
   // wait for hasura to be ready
   await waitForHasura();
-  tag('Wait for Hasura');
   // apply migrations and metadata
   await applyMigrations();
-  tag('Apply migrations');
-  // await applyMetadata();
-  await patchMetadata();
-  tag('Apply metadata');
+  await applyMetadata();
   // }
 
   // if (ENV.AUTH_SKIP_SERVE) {
@@ -45,7 +35,6 @@ export const start = async () => {
 
   app.listen(ENV.AUTH_PORT, () => {
     logger.info(`Running on port ${ENV.AUTH_PORT}`);
-    tag('Ready');
   });
   // }
 };
