+++ src/utils/index.ts
@@ -12,3 +12,4 @@ export * from './env';
 export * from './hasura';
 export * from './async-wrapper';
 export * from './webauthn';
+export * from './hasura-metadata';
