+++ test/routes/misc/custom-fields.test.ts
@@ -2,11 +2,10 @@ import { Client } from 'pg';
 import * as faker from 'faker';
 import { StatusCodes } from 'http-status-codes';
 
-import { reloadMetadata } from '@/metadata';
 import { ENV } from '@/utils';
 import { request } from '../../server';
 
-describe('metadata fields', () => {
+describe('user metadata field', () => {
   const firstName = faker.name.firstName();
   let client: Client;
 
@@ -15,7 +14,6 @@ describe('metadata fields', () => {
       connectionString: ENV.HASURA_GRAPHQL_DATABASE_URL,
     });
     await client.connect();
-    await reloadMetadata();
     await request.post('/change-env').send({
       AUTH_DISABLE_NEW_USERS: false,
       AUTH_EMAIL_SIGNIN_EMAIL_VERIFIED_REQUIRED: false,
@@ -23,7 +21,6 @@ describe('metadata fields', () => {
   });
 
   afterAll(async () => {
-    await reloadMetadata();
     await client.end();
   });
